% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_bglasso.R
\name{f_bglasso}
\alias{f_bglasso}
\title{Calls a Bayesian GLasso sampler}
\usage{
f_bglasso(X, r, q)
}
\arguments{
\item{X}{The input matrix}

\item{r}{The shape hyperparameter in a gamma distribution}

\item{q}{The rate hyperparameter in a gamma distribution}
}
\value{
One iterate of a Bayesian GLasso full conditional
}
\description{
This functions takes a matrix X and two numbers r and q as inputs and returns a draw from the full conditional of a Bayesian Graphical LASSO distribution.
}
\references{
Hao Wang.  Bayesian Graphical Lasso Models and Efficient Posterior Computation.  Bayesian Anal. 7(4):867-886.  December 2012
}
