% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_gwish.R
\name{f_gwish}
\alias{f_gwish}
\title{Calls a G-Wishart sampler}
\usage{
f_gwish(G, m, M)
}
\arguments{
\item{G}{An upper-triangular adjaceny matrix}

\item{m}{The degree of freedom in  a G-Wishart distribution}

\item{M}{The scale matrix in a G-Wishart distribution}
}
\value{
One draw from a W_G(m,M) distribution
}
\description{
This functions takes an adjaceny matrix G, a degree of freedom m, and a scale matrix M, and returns a draw form a W_G(m,M) distribution
}
