% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_gwish_cont.R
\name{f_gwish_cont}
\alias{f_gwish_cont}
\title{Calls a G-Wishart sampler}
\usage{
f_gwish_cont(X, bdgraph_obj, dfprior)
}
\arguments{
\item{X}{The data input matrix (D by p)}

\item{bdgraph_obj}{The output of a previous BDgraph G-wishart sampler}
}
\value{
One draw of a full conditional from a G-wishart prior in list form
}
\description{
This functions takes a matrix X and the output of a previous G-wishart prior sampler and returns a draw from a G-wishart prior in graph form.
}
