% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_iwish.R
\name{f_iwish}
\alias{f_iwish}
\title{Calls an Inverse-Wishart sampler}
\usage{
f_iwish(v, S)
}
\arguments{
\item{v}{The degrees of freedom in an IW(v,S) distribution}

\item{S}{The scale matrix in an IW(v,S) distribution}
}
\value{
A draw from an IW(v,S) distribution
}
\description{
This functions takes a double v and a scale matrix S to draw from an IW(v,S) distribution.  This is for uses in Cpp code.
}
