\name{LTNLDA-package}

\alias{LTNLDA-package}

\docType{package}

\title{
\packageTitle{LTNLDA}
}

\description{
\packageDescription{LTNLDA}
}

\details{
The DESCRIPTION file:
\packageDESCRIPTION{LTNLDA}
\packageIndices{LTNLDA}
}

\author{
\packageAuthor{LTNLDA}
Maintainer: \packageMaintainer{LTNLDA}
}

\keyword{ package }

