% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_Sampler.R
\name{LTNLDA}
\alias{LTNLDA}
\title{Runs an LTN-LDA Gibbs sampler.}
\usage{
LTNLDA(
  ps,
  K,
  threshold,
  iterations = 1000,
  burnin = 10000,
  thin = 10,
  alpha = 1,
  a1 = 10,
  a2 = 10^4,
  b = 10,
  Lambda = NULL
)
}
\arguments{
\item{ps}{A phyloseq object containing an otu_table() and a phy_tree() with an edge matrix.  That is, otu_table(ps) and phy_tree(ps)$edge both exist.}

\item{K}{An integer specifying the number of modeled subcommunities.}

\item{threshold}{An integer specifying the threshold controlling cross-sample heterogeneity.}

\item{iterations}{The number of iterations to record.  Default value is 1000.}

\item{burnin}{The number of burnin iterations to run before recording values.  The default value is 10000.}

\item{thin}{The amount by which we thin the chain.  A value of X means that 1 every X values is recorded after the burnin.  The default value is 10.}

\item{alpha}{A double specifying the prior on the subcommunity-sample proportions.  The default value is 1.}

\item{a1}{A double specifying the shape parameter for nodes with less descendants than threshold.  The default value is 10.}

\item{a2}{A double specifying the shape parameter for nodes with greater than or equal to descendants than threshold.  The default value is 10^4.}

\item{b}{A double specifying the rate parameter for all nodes.  The default value is 10.}

\item{Lambda}{A matrix specifying a covariance prior for the mu_k.  The default value is diag(V) where V is the number of leaves.}
}
\value{
A list with 11 entries.  Mean_Post_Phi_d contains the posterior mean estimate for the subcommunity-sample distributions phi_d. Mean_Post_Psi_pdk contains the posterior mean estimates for the log-odds psi_{p,d,k}. Mean_Post_Beta_kd contains the posterior mean estimate for the sample and subcommunity specific multinomial distributions beta_{k,d}. Mean_Post_Mu_k contains the posterior mean estimates for the average lod-odds mu_k. Mean_Post_Beta_k contains the posterior mean estimates for the average subcommunity multinomial distributions beta_k. Mean_Post_Sigma_ppk contains the posterior mean estimates for the covariance matrices Sigma_k. Chain_Phi contains the Markov Chains for the phi_d. Chain_Psi contains the Markov Chains for the psi_{p,d,k}. Chain_Mu contains the Markov Chains for the mu_k. Chain_Sigma contains the Markov Chains for the Sigma_k. Final_Iterate_Counts contains  the y_{d,k}(A) count data for the final iteration.
}
\description{
This function takes a phyloseq object, modeled, and modeled subcommunities and thresholds as inputs.  It runs a collapsed blocked Gibbs sampler for LTNLDA for these inputs, and returns a list containing posterior mean estimates and Markov Chains for various parameters, as well as the y_{d,k}(A) count data for the final iteration.  For a detailed example see the vignette "LTN-LDA".
}
\examples{
data("ps",package = "LTNLDA")
K = 2
threshold = 10
model = LTNLDA(ps,K,threshold)
}
\references{
ADD PAPER INFORMATION ONCE WE KNOW IT
}
