% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_Sampler_Block_Cov.R
\name{LTNLDA_block_cov}
\alias{LTNLDA_block_cov}
\title{Runs an LTN-LDA Gibbs sampler.  Experimental, trying out a new covariance structure.  Run at own risk.}
\usage{
LTNLDA_block_cov(
  ps,
  K,
  C = 5,
  iterations = 1000,
  burnin = 10000,
  thin = 10,
  alpha = 1,
  a_L = 10,
  b_L = 50,
  a_U = 10^4,
  b_U = 10,
  r = 100,
  q = 100,
  Lambda = NULL
)
}
\arguments{
\item{ps}{A phyloseq object containing an otu_table() and a phy_tree() with an edge matrix.  That is, otu_table(ps) and phy_tree(ps)$edge both exist.}

\item{K}{An integer specifying the number of modeled subcommunities.}

\item{C}{An integer specifying the threshold controlling cross-sample heterogeneity.  The default value is 5.  Using the default value may result in unreliable inference.}

\item{iterations}{The number of iterations to record.  Default value is 1000.}

\item{burnin}{The number of burnin iterations to run before recording values.  The default value is 10000.}

\item{thin}{The amount by which we thin the chain.  A value of X means that 1 every X values is recorded after the burnin.  The default value is 10.}

\item{alpha}{A double specifying the prior on the subcommunity-sample proportions.  The default value is 1.}

\item{a_L}{A double which is multipled by the scale matrix in the Inverse Wishart prior for the lower covariance matrix.  The default value is 10.}

\item{b_L}{A double which is multipled by the degrees of freedom in the Inverse Wishart prior for the lower covariance matrix.  The default value is 50.}

\item{a_U}{A double which is multipled by the scale matrix in the Inverse Wishart prior for the upeer covariance matrix.  The default value is 10^4.}

\item{b_U}{A double which is multipled by the degrees of freedom in the Inverse Wishart prior for the upper covariance matrix.  The default value is 10.}

\item{r}{A double serving as a hyperparameter for the Bayesian GLasso prior.  The default value is 1.}

\item{q}{A double serving as a hyperparameter for the Bayesian GLasso prior.  The default value is 1.}

\item{Lambda}{A matrix specifying a covariance prior for the mu_k.  The default value is diag(V) wher0e V is the number of leaves.}
}
\value{
A list with 8 entries.  
Mean_Post_Phi_d contains the posterior mean estimate for the subcommunity-sample distributions phi_d. 
Mean_Post_Beta_kd contains the posterior mean estimate for the sample and subcommunity specific multinomial distributions beta_{k,d}.
Mean_Post_Beta_k contains the posterior mean estimates for the average subcommunity multinomial distributions beta_k. 
Chain_Phi contains the Markov Chains for the phi_d. 
Chain_Psi contains the Markov Chains for the psi_{p,d,k}. 
Chain_Mu contains the Markov Chains for the mu_k. 
Chain_Sigma contains the Markov Chains for the Sigma_k. 
phyloseq contains the phyloseq object the Gibbs sampler ran on.
}
\description{
This function takes a phyloseq object and the modelled number of subcommunities as inputs. It runs a collapsed blocked Gibbs sampler for LTNLDA, and returns a list containing posterior mean estimates for some parameters, Markov Chains for all meaningful parameters, and the phyloseq object.  For a detailed example see the vignette "LTN-LDA".
}
\examples{
data("ps",package = "LTNLDA")
K = 2
model = LTNLDA(ps,K)
}
\references{
ADD PAPER INFORMATION ONCE WE KNOW IT
}
