% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gibbs_Sampler_Perp.R
\name{LTNLDA_Perplexity}
\alias{LTNLDA_Perplexity}
\title{Runs a modified LTN-LDA Gibbs sampler to estimate perplexity on a test set.}
\usage{
LTNLDA_Perplexity(
  model,
  ps,
  iterations = 1000,
  burnin = 10000,
  thin = 10,
  alpha = 1
)
}
\arguments{
\item{model}{The output of an LTNLDA model trained on a training set.}

\item{ps}{A phyloseq object corresponding to the test set containing an otu_table() and a phy_tree() with an edge matrix.  That is, otu_table(ps) and phy_tree(ps)$edge both exist.}

\item{iterations}{The number of iterations to record.  Default value is 1000.}

\item{burnin}{The number of burnin iterations to run before recording values.  The default value is 10000.}

\item{thin}{The amount by which we thin the chain.  A value of X means that 1 every X values is recorded after the burnin.  The default value is 10.}

\item{alpha}{A double specifying the prior on the subcommunity-sample proportions.  The default value is 1.}
}
\value{
A list with 5 entries. Perplexity is a double which is the estimated perplexity on the test set.  Mean_Post_Phi_d contains the posterior mean estimate for the subcommunity-sample distributions phi_d on each sample in the test set.  Mean_Post_Psi_pdk contains the posterior mean estimates for the log-odds psi_{p,d,k} for the test set. Chain_Phi contains the Markov Chains for the phi_d. Chain_Psi contains the Markov Chains for the psi_{p,d,k}.
}
\description{
This function takes the output of an LTNLDA function ran on the training set, an edge matrix specifying a tree (which should be the same tree as for the training set), and a matrix specifying counts of sequencing reads per sample on the test set. It runs a modified collapsed blocked Gibbs sampler for LTNLDA for these inputs to estimate the phi_{d} and psi_{p,d,k} on the test set, and returns a list containing a perplexity estimate on the test set as well as  posterior mean estimates and Markov Chains for the parameters.  For a detailed example see the vignette "Perplexity".
}
\examples{
perp = LTNLDA_Perplexity(model, test_ps)
}
\references{
ADD PAPER INFORMATION ONCE WE KNOW IT
}
