% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubAbund.R
\name{SubAbund}
\alias{SubAbund}
\title{Ranks subcommunities by order of average abundance.}
\usage{
SubAbund(model)
}
\arguments{
\item{model}{The output of an LTNLDA model trained on a training set.}
}
\value{
A list such that out[[k]] contains a matrix listing the top n ASVs in subcommunity k in descending order and with the corresponding prevalence.
}
\description{
This function takes the output of an LTNLDA function as an input, and finds the top n ASVs characterizing each subcommunity.
}
\examples{
data("ps",package = "LTNLDA")
K = 2
model = LTNLDA(ps,K)
SubAbund(model)
}
\references{
ADD PAPER INFORMATION ONCE WE KNOW IT
}
