% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{Summary}
\alias{Summary}
\title{Provides high level summary of model results.}
\usage{
Summary(model, n = 5)
}
\arguments{
\item{model}{The output of an LTNLDA model trained on a training set.}

\item{n}{We find the n most prevalent ASVs in each subcommunity, with proportions.}
}
\value{
Do later
}
\description{
This function takes the output of an LTNLDA function as an input, and finds the average subcommunity abundance in decreasing order, as well as the top n ASVs characterizing each subcommunity.  More detailed information in "LTN-LDA" vignette.
}
\examples{
data("ps",package = "LTNLDA")
K = 2
model = LTNLDA(ps,K)
Summary(model)
}
\references{
ADD PAPER INFORMATION ONCE WE KNOW IT
}
