% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n.R
\name{top_n}
\alias{top_n}
\title{Finds the n most prevalent ASVs in each subcommunity for a trained LTN-LDA model.}
\usage{
top_n(model, n = 5)
}
\arguments{
\item{model}{The output of an LTNLDA model trained on a training set.}

\item{n}{We find the n most prevalent ASVs in each subcommunity, with proportions.}
}
\value{
A list such that out[[k]] contains a matrix listing the top n ASVs in subcommunity k in descending order and with the corresponding prevalence.
}
\description{
This function takes the output of an LTNLDA function as an input, and finds the top n ASVs characterizing each subcommunity.
}
\examples{
data("ps",package = "LTNLDA")
K = 2
model = LTNLDA(ps,K)
top_n(model)
}
\references{
ADD PAPER INFORMATION ONCE WE KNOW IT
}
