% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_gwish.R
\name{f_gwish}
\alias{f_gwish}
\title{Calls a G-Wishart sampler}
\usage{
f_gwish(X, dfprior)
}
\arguments{
\item{X}{The data input matrix (D by p)}
}
\value{
One draw of a full conditional from a G-wishart prior in list form
}
\description{
This functions takes a matrix X and returns a draw from a G-wishart prior in graph form.
}
